Fix number parsing overruns.
When parsing numbers I just looked for all valid symbols that _could_ occur within a number, and assumed that json_get_number_size had verified the number was correct. This was wrong, because the way json.h is setup it will parse a number, and once it gets to the first character that it doesn't know about will bail successfully and then set about parsing (this allows us to support comments for instance).